refactor(heroRoller): move dice into new subsystem
