+++ module/testing/testing-dice.mjs
@@ -1,14 +1,14 @@
 import { setQuenchTimeout } from "./quench-helper.mjs";
 
-import { HeroRoller } from "../utility/dice.mjs";
+import { HeroRoller } from "../heroRoller/dice.mjs";
 
 import {
     Roll1Mock,
+    Roll1Through6Mock,
     Roll2Mock,
     Roll3Mock,
     Roll5Mock,
     Roll6Mock,
-    Roll1Through6Mock,
     RollAlternatingLuckAndUnluck,
 } from "./dice-testing-helper.mjs";
